refactor: migrate compareUrgencyFunction to items utils"
