+++ src/api/firebase.js
@@ -13,9 +13,6 @@ import {
 	transformToJSDate,
 	getDaysBetweenDates,
 	getNextPurchaseDate,
-	getItemDaysUntilNextPurchase,
-	getItemDaysSinceLastPurchase,
-	sortItems,
 } from '../utils';
 import { calculateEstimate } from '@the-collab-lab/shopping-list-utils';
 /**
@@ -125,20 +122,4 @@ export async function deleteItem(listId, item) {
 		console.log('Failed to delete item', error);
 		return { success: false, error: error.message };
 	}
-}
-
-export function comparePurchaseUrgency(data) {
-	const inactiveItems = sortItems(
-		data.filter((item) => {
-			return getItemDaysSinceLastPurchase(item) >= 60;
-		}),
-	);
-
-	const activeItems = sortItems(
-		data.filter((item) => {
-			return getItemDaysSinceLastPurchase(item) < 60;
-		}),
-	);
-
-	return [...activeItems, ...inactiveItems];
 }
